requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-emoji-picker==0.0.6
django-etc==1.3.0
django-hitcount==1.3.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
